requirements.txt
openai
python-dotenv
aioboto3
black
ruff